requirements.txt
pytesseract==0.3.10
Pillow==9.2.0
openai==0.27.0
opencv-python==4.6.0.66
flask
flask-cors
python-dotenv
flask_jwt_extended
python-dateutil
psycopg2-binary
pypika
phonenumbers
reportlab
requests
